blockchain implementation that uses a proof-of-work consensus algorithm to prevent tampering
the is_valid_proof function is used to ensure that the hash of a new block meets certain criteria, which makes it difficult for an attacker to alter the blockchain. The attacker would need to recalculate the hash of the altered block and all subsequent blocks, which would require a significant amount of computational power.

The is_valid_proof function uses a proof-of-work consensus algorithm, which requires the attacker to perform a large number of calculations to find a hash that meets the criteria. This makes it computationally expensive to alter the blockchain, which increases the security of the system.

The create_new_block function uses the is_valid_proof function to ensure that the hash of a new block meets the criteria before it is added to the blockchain. This helps to prevent tampering and ensures the integrity of the data stored in the blockchain.